+++ src/timedb.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-int timedbSearch(const void* data, size_t size, uint32_t* framesArray, int framesArraySize);
+int timedbSearch(const void* data, uint32_t size, uint32_t* framesArray, int framesArraySize);
